remove unnecessary casting

+++ 3/include/solution.playlist.h
@@ -47,6 +47,6 @@ class Solution {
       }
     }
 
-    return (int)maxPlaylistLength;
+    return maxPlaylistLength;
   }
 };
